Change return value from int to double

+++ weather/WeatherStationSimulatio n/Program.cs
@@ -23,7 +23,7 @@ class Program
 
 		static double CalculateAverage(int[] temperature)
 		{
-			int sum = 0;
+			double sum = 0;
 			for (int i = 0; i < temperature.Length; i++)
 			{
 				sum += temperature[i];
